+++ src/client/GraphQLRequestError.ts
@@ -5,15 +5,18 @@ export class GraphQLRequestError<
 > extends Error {
   public readonly response: Response;
   public readonly result?: ExecutionResult<TData>;
+  public readonly json?: unknown;
 
   constructor(
     message: string,
     response: Response,
-    result?: ExecutionResult<TData>
+    result?: ExecutionResult<TData>,
+    json?: unknown
   ) {
     super(message);
 
     this.response = response;
     this.result = result;
+    this.json = result ?? json;
   }
 }